requirements.txt
amqp==5.0.7; python_version >= "3.7"
archiveis==0.0.9
asgiref==3.4.1; python_version >= "3.8"
attrs==21.2.0; python_full_version >= "3.6.7" and python_version >= "3.6"
autobahn==21.11.1; python_version >= "3.7"
automat==20.2.0; python_full_version >= "3.6.7" and python_version >= "3.6"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
billiard==3.6.4.0; python_version >= "3.7"
bs4==0.0.1
cachecontrol==0.12.10; python_version >= "3.6"
celery==5.2.1; python_version >= "3.7"
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0
charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3.6"
click-didyoumean==0.3.0; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.7"
click-plugins==1.1.1; python_version >= "3.7"
click-repl==0.2.0; python_version >= "3.7"
click==8.0.3; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
constantly==15.1.0; python_full_version >= "3.6.7" and python_version >= "3.6"
cryptography==36.0.0; python_full_version >= "3.6.7" and python_version >= "3.7"
daphne==3.0.2; python_version >= "3.6"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
django-celery-beat==2.1.0
django-celerybeat-status==0.0.10
django-debug-toolbar==3.2.3; python_version >= "3.6"
django-redis==4.12.1; python_version >= "3.5"
django-timezone-field==4.2.1; python_version >= "3.5"
django==4.0a1; python_version >= "3.8"
djangorestframework==3.12.4; python_version >= "3.5"
flower==1.0.0
gevent==21.12.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_version > "3.5")
greenlet==1.1.2; python_version >= "2.7" and python_full_version < "3.0.0" and platform_python_implementation == "CPython" or python_version > "3.5" and python_full_version < "3.0.0" and platform_python_implementation == "CPython" or python_version > "3.5" and platform_python_implementation == "CPython" and python_full_version >= "3.5.0"
humanize==3.13.1; python_version >= "3.6"
hyperlink==21.0.0; python_full_version >= "3.6.7" and python_version >= "3.7"
idna==3.3; python_full_version >= "3.6.7" and python_version >= "3.6" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7")
importlib-metadata==4.8.2; python_version < "3.10" and python_version >= "3.6"
incremental==21.3.0; python_full_version >= "3.6.7" and python_version >= "3.6"
kombu==5.2.2; python_version >= "3.7"
lxml==4.7.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
markdown==3.3.6; python_version >= "3.6"
msgpack==1.0.3; python_version >= "3.6"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
praw==7.5.0; python_version >= "3.6" and python_version < "4.0"
prawcore==2.3.0; python_version >= "3.6" and python_version < "4.0"
prometheus-client==0.12.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
prompt-toolkit==3.0.24; python_full_version >= "3.6.2" and python_version >= "3.7"
psycopg2-binary==2.9.2; python_version >= "3.6"
pyasn1-modules==0.2.8; python_full_version >= "3.6.7" and python_version >= "3.6"
pyasn1==0.4.8; python_full_version >= "3.6.7" and python_version >= "3.6"
pycparser==2.21; python_version >= "2.7" and python_full_version < "3.0.0" and platform_python_implementation == "CPython" and sys_platform == "win32" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_version > "3.5") or platform_python_implementation == "CPython" and sys_platform == "win32" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_version > "3.5") and python_full_version >= "3.4.0"
pyopenssl==21.0.0; python_full_version >= "3.6.7" and python_version >= "3.6"
python-crontab==2.6.0
python-dateutil==2.8.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
pytz==2021.3; python_version >= "3.7"
redis==4.0.2; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
sentry-sdk==1.5.0
service-identity==21.1.0; python_full_version >= "3.6.7" and python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.7" and python_version >= "3.7"
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
sqlparse==0.4.2; python_version >= "3.8"
tornado==6.1; python_version >= "3.5"
tweepy==4.4.0; python_version >= "3.6"
twisted-iocpsupport==1.0.2; python_full_version >= "3.6.7" and python_version >= "3.6" and platform_system == "Windows"
twisted==21.7.0; python_full_version >= "3.6.7" and python_version >= "3.6"
txaio==21.2.1; python_version >= "3.7"
typing-extensions==4.0.1; python_full_version >= "3.6.7" and python_version >= "3.6"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
update-checker==0.18.0; python_version >= "3.6" and python_version < "4.0"
urllib3==1.26.7; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
vine==5.0.0; python_version >= "3.7"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.7"
websocket-client==1.2.3; python_version >= "3.6" and python_version < "4.0"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
wrapt==1.13.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
zipp==3.6.0; python_version < "3.10" and python_version >= "3.6"
zope.event==4.5.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_version > "3.5"
zope.interface==5.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
zstandard==0.16.0; python_version >= "3.6"
